requirements.txt
numpy==1.22.2
pandas==1.4.0
Pillow==9.1.0
plotly==5.7.0
streamlit==1.9.0
